backend/requirements.txt
Flask
pdfminer.six
spacy
scikit-learn
nltk
